backend/requirements.txt
Flask==3.0.3
google-cloud-storage==1.42.0
gunicorn==20.1.0
Werkzeug==3.0.3
python-dotenv
flask-cors
pdfquery
pdfminer.six
temp
transformers
torch 
numpy
pinecone-client
psycopg2-binary
celery[sqlalchemy]
